Signin form errors (#28)
* Display errors when signing in

* Create a reusable composable to handle form errors
- include the composables directory in vite.config.ts to
enable any composables we create to automatically be imported

* Provide real time validation on the login form
- expand the composable to add a realTimeErrors state to dynamically
handle validation for email and password fields
- create a utility function to validate inputs and trigger
these validations on user input

* Fix UI issue with error message
- The error message was being displayed continuously
whilst keying data in the fields making for an unpleasant
user experience. Use the debouncedWatch package from VueUse to
delay triggering form validation for 1 second after the user starts
typing.

* Maintenance: provide a type for the realtimeErrors ref

Map over the loginForm object to dynamically map types
for the errors ref which can be extended to use for the
registerForm as well.